De-duplicat _push_to_os.
